refactor: PendingMembersHeader 컴포넌트로 헤더 컴포넌트 분리
